+++ test/localregex_test.dart
@@ -4,134 +4,162 @@ import 'package:localregex/localregex.dart';
 void main() {
   final localregex = LocalRegex();
   test('Validate email address', () {
-    try {
-      assert(localregex.isEmail("ngmangudya@gmail.com"), true);
-    } on AssertionError catch (e) {
-      print(e);
-    }
+    // Act
+    bool isValid = localregex.isEmail("ngmangudya@gmail.com");
+    // Assert
+    expect(isValid, true);
   });
 
   group("Test +263 mobile numbers", () {
+    test(":Validate netone number with +26371", () {
+      // Act
+      bool isValid = localregex.isNetone("+263713700601");
+      // Assert
+      expect(isValid, true);
+    });
+    test(":Validate telecel number with +26373", () {
+      // Act
+      bool isValid = localregex.isTelecel("+263733220033");
+      // Assert
+      expect(isValid, true);
+    });
+
+    test(":Validate econet number +26377", () {
+      // Act
+      bool isValid = localregex.isEconet("+263777213388");
+      // Assert
+      expect(isValid, true);
+    });
+
+    test(":Validate econet number +26378", () {
+      // Act
+      bool isValid = localregex.isEconet("+263782213388");
+      // Assert
+      expect(isValid, true);
+    });
+  });
+
+  group("Test 263 mobile numbers", () {
     test(":Validate netone number", () {
-      try {
-        assert(localregex.isNetone("263713700601"), true);
-      } on AssertionError catch (e) {
-        print(e);
-      }
+      // Act
+      bool isValid = localregex.isNetone("263713700601");
+      // Assert
+      expect(isValid, true);
     });
     test(":Validate telecel number", () {
-      try {
-        assert(localregex.isTelecel("2630733220033"), true);
-      } on AssertionError catch (e) {
-        print(e);
-      }
+      // Act
+      bool isValid = localregex.isTelecel("263733220033");
+      // Assert
+      expect(isValid, true);
     });
 
     test(":Validate econet number 26377", () {
-      try {
-        assert(localregex.isEconet("263777213388"), true);
-      } on AssertionError catch (e) {
-        print(e);
-      }
+      // Act
+      bool isValid = localregex.isEconet("263777213388");
+      // Assert
+      expect(isValid, true);
     });
 
     test(":Validate econet number 26378", () {
-      try {
-        assert(localregex.isEconet("+263782213388"), true);
-      } on AssertionError catch (e) {
-        print(e);
-      }
+      // Act
+      bool isValid = localregex.isEconet("263782213388");
+      // Assert
+      expect(isValid, true);
     });
   });
 
   group("Test 07 mobile numbers", () {
     test(":Validate netone number", () {
-      try {
-        assert(localregex.isNetone("0713700601"), true);
-      } on AssertionError catch (e) {
-        print(e);
-      }
+      // Act
+      bool isValid = localregex.isNetone("0713700601");
+      // Assert
+      expect(isValid, true);
     });
     test(":Validate telecel number", () {
-      try {
-        assert(localregex.isTelecel("0733220033"), true);
-      } on AssertionError catch (e) {
-        print(e);
-      }
+      // Act
+      bool isValid = localregex.isTelecel("0733220033");
+      // Assert
+      expect(isValid, true);
     });
 
     test(":Validate econet number 077", () {
-      try {
-        assert(localregex.isEconet("0777213388"), true);
-      } on AssertionError catch (e) {
-        print(e);
-      }
+      // Act
+      bool isValid = localregex.isEconet("0777213388");
+      // Assert
+      expect(isValid, true);
     });
 
     test(":Validate econet number 078", () {
-      try {
-        assert(localregex.isEconet("0782213388"), true);
-      } on AssertionError catch (e) {
-        print(e);
-      }
+      // Act
+      bool isValid = localregex.isEconet("0782213388");
+      // Assert
+      expect(isValid, true);
     });
   });
 
   test("Validate driver's license number", () {
-    try {
-      assert(localregex.isZwDriversLicence("98336DM"), true);
-    } on AssertionError catch (e) {
-      print(e);
-    }
+    // Act
+    bool isValid = localregex.isZwDriversLicence("98336DM");
+    // Assert
+    expect(isValid, true);
   });
 
   test("Validate passport number", () {
-    try {
-      assert(localregex.isZwPassport("FN633423"), true);
-    } on AssertionError catch (e) {
-      print(e);
-    }
+    // Act
+    bool isValid = localregex.isZwPassport("FN633423");
+    // Assert
+    expect(isValid, true);
   });
 
   test("Validate number plate", () {
-    try {
-      assert(localregex.isZwNumberPlate("ADW 4746"), true);
-    } on AssertionError catch (e) {
-      print(e);
-    }
+    // Act
+    bool isValid = localregex.isZwNumberPlate("ADW 4746");
+    // Assert
+    expect(isValid, true);
   });
 
   group("Test national id", () {
     test(":Validate national id without a space", () {
-      try {
-        assert(localregex.isZw("75-511218 K 44"), true);
-      } on AssertionError catch (e) {
-        print(e);
-      }
+      // Act
+      bool isValid = localregex.isZimbabwean("75-511218 K 44");
+      // Assert
+      expect(isValid, true);
     });
     test(":Validate national id with a space", () {
-      try {
-        assert(localregex.isZw("75 511218 K 44"), true);
-      } on AssertionError catch (e) {
-        print(e);
-      }
+      // Act
+      bool isValid = localregex.isZimbabwean("75 511218 K 44");
+      // Assert
+      expect(isValid, true);
     });
   });
 
   group("Testing if mobile numbers belongs to Econet or Netone or Telecel", () {
     test(":Validate mobile numbers", () {
-      try {
-        assert(localregex.isZwMobile("0777213388"), true);
-        assert(localregex.isZwMobile("+263777213388"), true);
-        assert(localregex.isZwMobile("0713700601"), true);
-        assert(localregex.isZwMobile("+263713700601"), true);
-        assert(localregex.isZwMobile("0733220033"), true);
-        assert(localregex.isZwMobile("+263733220033"), true);
-        assert(localregex.isZwMobile("0782213388"), true);
-        assert(localregex.isZwMobile("+263782213388"), true);
-      } on AssertionError catch (e) {
-        print(e);
-      }
+      bool isValid;
+
+      isValid = localregex.isZwMobile("0777213388");
+      expect(isValid, true);
+      isValid = localregex.isZwMobile("+263777213388");
+      expect(isValid, true);
+      isValid = localregex.isZwMobile("0713700601");
+      expect(isValid, true);
+      isValid = localregex.isZwMobile("+263713700601");
+      expect(isValid, true);
+      isValid = localregex.isZwMobile("0733220033");
+      expect(isValid, true);
+      isValid = localregex.isZwMobile("+263733220033");
+      expect(isValid, true);
+      isValid = localregex.isZwMobile("0782213388");
+      expect(isValid, true);
+      isValid = localregex.isZwMobile("+263782213388");
+      expect(isValid, true);
+      isValid = localregex.isZwMobile("263777213388");
+      expect(isValid, true);
+      isValid = localregex.isZwMobile("263782213388");
+      expect(isValid, true);
+      isValid = localregex.isZwMobile("263713700601");
+      expect(isValid, true);
+      isValid = localregex.isZwMobile("263733220033");
     });
   });
 }